Moved some tests around, added test_base_data file, added docstrings to tests.
