+++ matplotcheck/tests/test_base.py
@@ -1,10 +1,6 @@
 """Tests for the base module"""
 import pytest
 import matplotlib.pyplot as plt
-import numpy as np
-import random
-
-import pdb
 
 
 def test_line_plot(pt_line_plt):
@@ -55,78 +51,9 @@ def test_options(pt_line_plt):
     plt.close()
 
 
-def test_correct_title(pt_line_plt):
-    """Check that the correct plot title is grabbed from the axis object.
-    Note that get_titles maintains case."""
-
-    assert "Plot Title" in pt_line_plt.get_titles()[1]
-    plt.close()
-
-
-"""DATACHECK TESTS"""
-
-
-def test_assert_xydata_scatter(pt_scatter_plt, pd_df):
-    """Checks points in scatter plot against expected data"""
-    pt_scatter_plt.assert_xydata(pd_df, xcol="A", ycol="B")
-    plt.close()
-
-
-def test_assert_xydata_tolerance(pt_scatter_plt, pd_df):
-
-    for i in range(len(pd_df["A"])):
-        pd_df["A"][i] = pd_df["A"][i] + (np.floor(pd_df["A"][i] * 0.25))
-        pd_df["B"][i] = pd_df["B"][i] + (np.floor(pd_df["B"][i] * 0.25))
-    pt_scatter_plt.assert_xydata(pd_df, xcol="A", ycol="B", tolerence=0.5)
-    plt.close()
-
-
-def test_assert_xydata_tolerance_fail(pt_scatter_plt, pd_df):
-    pd_df["A"][1] = pd_df["A"][1] * 2
-    with pytest.raises(AssertionError, match="Incorrect data values"):
-        pt_scatter_plt.assert_xydata(pd_df, xcol="A", ycol="B", tolerence=0.1)
-    plt.close()
-
-
-def test_assert_xydata_changed_data(pt_scatter_plt, pd_df):
-    """assert_xydata should fail when we change the data"""
-    pd_df["B"][1] += 5
-    with pytest.raises(AssertionError, match="Incorrect data values"):
-        pt_scatter_plt.assert_xydata(pd_df, xcol="A", ycol="B")
-    plt.close()
-
-
-def test_assert_xydata_scatter_points_only(pt_scatter_plt, pd_df):
-    """Checks points in scatter plot against expected data"""
-    pt_scatter_plt.assert_xydata(pd_df, xcol="A", ycol="B", points_only=True)
-    plt.close()
-
-
-def test_assert_xydata_changed_data_points_only(pt_scatter_plt, pd_df):
-    """assert_xydata should fail when we change the data"""
-    pd_df["B"][1] += 5
-    with pytest.raises(AssertionError, match="Incorrect data values"):
-        pt_scatter_plt.assert_xydata(
-            pd_df, xcol="A", ycol="B", points_only=True
-        )
-    plt.close()
-
-
+"""
+Test fo geopandas. Commenting out until later.
 def test_assert_xydata_geo(pd_gdf, pt_geo_plot):
     pt_geo_plot.assert_xydata(pd_gdf)
     plt.close()
-
-
-'''
-def test_assert_xydata_timeseries(pt_time_line_plt, pd_df_timeseries):
-    """Commenting this out for now as this requires a time series data object
-    this is failing because the time data needs to be in seconds like how
-    mpl saves it. """
-    pt_time_line_plt.assert_xydata(pd_df_timeseries, xcol='time', ycol='A', xtime=True)
-'''
-
-
-def test_assert_xydata_xlabel(pt_bar_plt, pd_df):
-    pd_df["A"] = pd_df["A"].apply(str)
-    pt_bar_plt.assert_xlabel_ydata(pd_df, xcol="A", ycol="B")
-    plt.close()
+"""
